Moved functions that may be used by convert_treefrog to a common file.
